Init commit refactoring utils package
