+++ utils/models.py
@@ -1,4 +1,28 @@
 from fastai.callback.core import Callback
+from fastai.vision.learner import create_cnn_model
+
+try:
+    from efficientnet_pytorch import EfficientNet
+except ImportError:
+    print('EfficientNet models cannot be used because `efficientnet_pytorch` package is not installed')
+
+
+def create_model(model_arq, n_out, model=None, pretrained=True, n_in=1, ema=False, bn_final=False):
+    if model is None:
+      if type(model_arq) is str and model_arq[:12].lower == 'efficientnet':
+        model = EfficientNet.from_pretrained(model_arq, num_classes=n_out, include_top=True, in_channels=n_in)
+      else:
+        model = create_cnn_model(model_arq, n_out=n_out, cut=None, pretrained=pretrained, n_in=n_in, bn_final=bn_final)
+    
+    if torch.cuda.is_available():
+      model = model.cuda()
+
+    if ema:
+        for param in model.parameters():
+            param.detach_()
+
+    return model
+
 
 class EMAModel(Callback):
     """ Evaluate model during validation with a EMA model
