+++ utils/organize.py
@@ -12,183 +12,6 @@ import pandas as pd
 
 import pydicom
 
-import torch
-
-from fastai.data.transforms import IndexSplitter
-from fastai.basics import mask2idxs
-
-from fastai.vision.learner import create_cnn_model
-from efficientnet_pytorch import EfficientNet
-
-def categorical_to_one_hot(x, n_out):
-    """ Transform categorical tensor to one hot encoding """
-    zeros = torch.zeros(len(x), n_out)
-    if torch.cuda.is_available():
-        zeros = zeros.cuda()
-
-    return zeros.scatter_(1, x.view(-1,1).long(), 1)
-
-def TestColSplitter(col='Dataset'):
-    "Split `items` (supposed to be a dataframe) by value in `col`"
-    def _inner(o):
-        assert isinstance(o, pd.DataFrame), "ColSplitter only works when your items are a pandas DataFrame"
-        train_idx = (o.iloc[:,col] if isinstance(col, int) else o[col]) == 'train'
-        valid_idx = (o.iloc[:,col] if isinstance(col, int) else o[col]) == 'valid'
-        test_idx = (o.iloc[:,col] if isinstance(col, int) else o[col]) == 'test'
-        return IndexSplitter(mask2idxs(train_idx))(o)[1], IndexSplitter(mask2idxs(valid_idx))(o)[1], IndexSplitter(mask2idxs(test_idx))(o)[1]
-    return _inner
-
-
-def create_model(model_arq, n_out, model=None, pretrained=True, n_in=1, ema=False, bn_final=False):
-    if model is None:
-      if type(model_arq) is str and model_arq[:12].lower == 'efficientnet':
-        model = EfficientNet.from_pretrained(model_arq, num_classes=n_out, include_top=True, in_channels=n_in)
-      else:
-        model = create_cnn_model(model_arq, n_out=n_out, cut=None, pretrained=pretrained, n_in=n_in, bn_final=bn_final)
-    
-    if torch.cuda.is_available():
-      model = model.cuda()
-
-    if ema:
-        for param in model.parameters():
-            param.detach_()
-
-    return model
-
-
-def seed_everything(use_seed=0):
-    seed = use_seed if use_seed else random.randint(1,1000000)
-    print(f"Using seed: {seed}")
-
-    # python RNG
-    random.seed(seed)
-    os.environ['PYTHONHASHSEED'] = str(seed)
-
-    # pytorch RNGs
-    torch.manual_seed(seed)
-    if torch.cuda.is_available(): torch.cuda.manual_seed_all(seed)
-    if use_seed:
-        torch.backends.cudnn.deterministic = True
-        torch.backends.cudnn.benchmark     = False
-
-    # numpy RNG
-    np.random.seed(seed)
-
-    return seed
-
-
-def filter_fnames(
-    fnames,
-    metadata_raw_path,
-    check_DICOM_dict={
-        'Modality': ['CR', 'DR', 'DX'],
-    }
-):
-    """ Filter all the filenames which are fulfill the metadata conditions passed on `check_DICOM_dict` """
-
-    # Load metadata 
-    metadata_df = pd.read_csv(metadata_raw_path)
-
-    # Filter by the ones taht fulfill the conditions
-    metadata_df = df_check_DICOM(metadata_df, check_DICOM_dict)
-
-    # Create a DataFrame to compare filenames
-    check_df = pd.DataFrame(metadata_df.fname.str.split('/').str[-1], index=metadata_df.fname.str.split('/').str[-1])
-
-    # Loop over all the filenames
-    filter_fnames = []
-    for fname in tqdm(fnames):
-        try:
-            filename, ext = os.path.splitext(fname.name)
-        except AttributeError:
-            fname = Path(fname)
-            filename, ext = os.path.splitext(fname.name)
-
-        # Take into account the ones which seems to have extension ".PACSXXX" is not an extension
-        if ext.startswith('.PACS'):
-            filename = fname.name
-
-        # Check if is in the list and add it
-        try:
-            check_df.loc[filename, :]
-            filter_fnames.append(str(fname).replace(os.sep, '/'))
-        except KeyError:
-            continue
-    
-    return filter_fnames
-
-
-def default_check_DICOM_dict():
-    """ Get default values for check DICOM dictionary """
-
-    check_DICOM_dict = {
-        'SeriesDescription': ['RODILLA AP', 'RODILLAS AP'],
-        'BodyPartExamined': ['LOWER LIMB', 'KNEE']
-    }
-
-    return check_DICOM_dict
-
-
-def df_check_DICOM(df, check_DICOM_dict):
-    """ Filter DataFrame on the rows that match the filter specified on `check_DICOM_dict` """
-    match = True
-    for key, value in check_DICOM_dict.items():
-        # Lambda function checking if DICOM file fulfills the condition
-        if key == 'function':
-            match = (match) & df.apply(value, axis=1)
-        else:
-            match = (match) & df[key].isin(value)
-    
-    return df[match]
-
-
-def sample_df_check_DICOM(df, check_DICOM_dict, max_samples_per_case=5):
-    """ Get a random sample of the filtered DataFrame determined by `check_DICOM_dict` appearing all the cases there """
-
-    df_match = df_check_DICOM(df, check_DICOM_dict)
-    
-    all_keys = list(check_DICOM_dict.keys())
-
-    return get_each_case_samples(df_match, all_keys, max_samples_per_case)
-
-
-def get_each_case_samples(df, all_keys, max_samples_per_case=5):
-    """ Extract all the different cases on `all_keys` and creating a DataFrame with the number set for all the cases """
-    cases_df = df[all_keys].drop_duplicates()
-    
-    concat_list = []
-    for i in range(len(cases_df.index)):
-        concat_list.append(
-            df[
-                (
-                    df[all_keys] == cases_df.iloc[i][all_keys]
-                ).all(axis=1)
-            ].sample(max_samples_per_case, replace=True)
-        )
-
-    return pd.concat(concat_list)
-
-
-def check_DICOM(dcm, check_DICOM_dict=None, debug=False):
-    """ Check the DICOM file if it is has the feature required """
-
-    if check_DICOM_dict is None:
-        check_DICOM_dict = default_check_DICOM_dict()
-
-    check = True
-    for key, value in check_DICOM_dict.items():
-        # Lambda function checking if DICOM file fulfills the condition
-        if key == 'function':
-            check = value(dcm)
-            if not check:
-                break
-        elif dcm.get(key) not in value:
-            check = False
-            if debug:
-                print(f'{key}: {dcm.get(key)} on Accession Number: {dcm.AccessionNumber}')
-            break
-    return check
-
 
 def check_metadata_label(raw_path, metadata_labels, label='ap'):
     """ Check if the path matches with the desired label on metadata_labels """
